+++ sdk/servicebus/azure-messaging-servicebus/src/test/java/com/azure/messaging/servicebus/TestUtils.java
@@ -64,6 +64,7 @@ public class TestUtils {
     static final int USE_CASE_EMPTY_ENTITY = 18;
     static final int USE_CASE_CANCEL_MESSAGES = 19;
     static final int USE_CASE_AUTO_COMPLETE = 20;
+    static final int USE_CASE_AMQP_TYPES = 21;
 
     // An application property key to identify where in the stream this message was created.
     static final String MESSAGE_POSITION_ID = "message-position";